criação componentes layout e dashboard e remoção código do app component
