+++ src/app/app.component.html
@@ -1,37 +1,1 @@
-<div class="container" [class.is-mobile]="isMobile()">
-  <mat-toolbar class="toolbar">
-    <button mat-icon-button (click)="snav.toggle()">
-      <mat-icon>menu</mat-icon>
-    </button>
-    <h1 class="app-name">Angular Web App</h1>
-  </mat-toolbar>
-
-  <mat-sidenav-container
-    class="sidenav-container"
-    [style.marginTop.px]="isMobile() ? 56 : 0"
-  >
-    <mat-sidenav
-      #snav
-      [mode]="isMobile() ? 'over' : 'side'"
-      [fixedInViewport]="isMobile()"
-      fixedTopGap="56"
-      [opened]="isMobile() ? false : true"
-    >
-      <mat-nav-list>
-        @for (nav of fillerNav; track nav) {
-        <a mat-list-item [routerLink]="nav.route">
-          <div class="flex-container">
-            <mat-icon matListIcon>{{ nav.icon }}</mat-icon>
-            {{ nav.name }}
-          </div>
-          <mat-divider></mat-divider>
-        </a>
-        }
-      </mat-nav-list>
-    </mat-sidenav>
-
-    <mat-sidenav-content>
-      <router-outlet />
-    </mat-sidenav-content>
-  </mat-sidenav-container>
-</div>
+<layout></layout>
